ENH: Refactor deps into subpackages
