+++ tests/conftest.py
@@ -90,7 +90,7 @@ def session_manager() -> Generator[SessionManager]:
     """Mocks the session manager and returns its replacement."""
     session_manager = SessionManager()
     with (
-        patch("fmu_settings_api.deps.session_manager", session_manager),
+        patch("fmu_settings_api.deps.session.session_manager", session_manager),
         patch("fmu_settings_api.session.session_manager", session_manager),
         patch("fmu_settings_api.v1.routes.session.session_manager", session_manager),
     ):
